Move context traversal stuff to Translate package
