+++ library/Core/package.yaml
@@ -24,8 +24,6 @@ dependencies:
 - containers
 - unordered-containers
 - lens
-- dlist
-- syb
 
 default-extensions:
   - NoImplicitPrelude
