+++ src/Juvix/Pipeline.hs
@@ -4,6 +4,7 @@ module Juvix.Pipeline where
 
 import qualified Juvix.Core as Core
 import qualified Juvix.Core.Common.Context as Context
+import qualified Juvix.Core.Common.Context.Traverse as Context
 import qualified Juvix.Core.FromFrontend as FF
 import qualified Juvix.Core.IR.Types as IR
 import qualified Juvix.Core.Parameterisation as P
